requirements.txt
Babel==0.9.6
backports.ssl-match-hostname==3.5.0.1
cffi==1.14.1
chardet==2.2.1
Cheetah==2.4.4
click==7.1.2
cloud-init==17.1
configobj==4.7.2
cryptography==3.0
decorator==3.4.0
enum34==1.0.4
Flask==1.0.2
futures==3.1.1
gevent==20.6.2
greenlet==0.4.16
idna==2.4
iniparse==0.4
ipaddress==1.0.16
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.2
jsonpointer==1.9
jsonschema==2.5.1
jwcrypto==0.4.2
kitchen==1.1.1
langtable==0.0.31
Markdown==2.4.1
MarkupSafe==1.1.1
oauthlib==2.0.1
perf==0.1
Pillow==2.0.0
ply==3.4
prettytable==0.7.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.0
Pygments==1.4
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-augeas==0.5.0
python-linux-procfs==0.4.9
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
redis==3.5.3
repoze.lru==0.4
requests==2.6.0
schedutils==0.4
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
tencentcloud-sdk-python==3.0.228
urlgrabber==3.10
urllib3==1.10.2
Werkzeug==0.16.0
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
zope.event==4.4
zope.interface==5.1.0
